♻️ Refactor: 상품 불필요 코드 제거&삭제 수정

See merge request s12-webmobile1-sub1/S12P11C108!82

+++ spring-project/src/main/java/com/c108/springproject/order/dto/response/OrderDetailResDto.java
@@ -16,10 +16,10 @@ import java.util.stream.Collectors;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderDetailResDto {
-    private int orderNo;
+    private Long orderNo;
     private int userNo;
     private int addressNo;
-    private BigInteger userCouponNo;
+    private Long userCouponNo;
     private String payment;
     private int totalPrice;
     private String requests;
