+++ spring-project/src/main/java/com/c108/springproject/recipe/dto/request/RecipeCreateReqDto.java
@@ -13,7 +13,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class RecipeCreateReqDto {
     private int categoryNo;
-    private BigInteger imageNo;
+    private Long imageNo;
     private String name;
     private int time;
     private int createdUser;
