+++ spring-project/src/main/java/com/c108/springproject/recipe/dto/response/RecipeResDto.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class RecipeResDto {
     private int recipeNo;
-    private BigInteger imageNo;
+    private Long imageNo;
     private int categoryNo;
     private String categoryName;
     private String name;
