+++ spring-project/src/main/java/com/c108/springproject/review/dto/request/ReviewUpdateReqDto.java
@@ -12,7 +12,7 @@ import java.math.BigInteger;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReviewUpdateReqDto {
-    private BigInteger imageNo;
+    private Long imageNo;
     private String content;
     private int rating;
     private int updatedUser;
